Fixed minor warnings caused by previous commit.

+++ oficonv/libsrc/citrus_viqr.c
@@ -297,7 +297,7 @@ _citrus_VIQR_mbrtowc_priv(_VIQREncodingInfo * ei,
     }
     if (ch == ESCAPE && m != ei->mroot)
         ++i;
-    psenc->chlen -= i;
+    psenc->chlen -= (int)i;
     memmove(&psenc->ch[0], &psenc->ch[i], psenc->chlen);
     wc = (m == ei->mroot) ? (_citrus_wc_t)ch : m->value;
     if (pwc != NULL)
